v1.0 design finished
This is the first release candidate for v1.0. I've changed the units
around so that each one always has just one drawer, and different units
may be made arbitrarily in sizes that are multiples of some unit length,
which in this case is 32mm. The way the units are designed, 2x2 units
may be arbitrarily combined with 1x1, 4x1, whatever, to make a fully
custom shelving unit.